refactor(iota-framework/iota-types): reduce amount of versioned types V2 to V1
