+++ crates/iota-types/src/gas.rs
@@ -19,7 +19,7 @@ pub mod checked {
         effects::{TransactionEffects, TransactionEffectsAPI},
         error::{ExecutionError, IotaResult, UserInputError, UserInputResult},
         gas_model::{
-            gas_predicates::gas_price_too_high, gas_v2::IotaGasStatus as IotaGasStatusV2,
+            gas_predicates::gas_price_too_high, gas_v2::IotaGasStatus as IotaGasStatusV1,
             tables::GasStatus,
         },
         iota_serde::{BigInt, Readable},
@@ -57,9 +57,7 @@ pub mod checked {
     #[enum_dispatch(IotaGasStatusAPI)]
     #[derive(Debug)]
     pub enum IotaGasStatus {
-        // V1 does not exists any longer as it was a pre mainnet version.
-        // So we start the enum from V2
-        V2(IotaGasStatusV2),
+        V1(IotaGasStatusV1),
     }
 
     impl IotaGasStatus {
@@ -88,7 +86,7 @@ pub mod checked {
                 .into());
             }
 
-            Ok(Self::V2(IotaGasStatusV2::new_with_budget(
+            Ok(Self::V1(IotaGasStatusV1::new_with_budget(
                 gas_budget,
                 gas_price,
                 reference_gas_price,
@@ -97,7 +95,7 @@ pub mod checked {
         }
 
         pub fn new_unmetered() -> Self {
-            Self::V2(IotaGasStatusV2::new_unmetered())
+            Self::V1(IotaGasStatusV1::new_unmetered())
         }
 
         // This is the only public API on IotaGasStatus, all other gas related
@@ -108,7 +106,7 @@ pub mod checked {
             gas_budget: u64,
         ) -> UserInputResult {
             match self {
-                Self::V2(status) => status.check_gas_balance(gas_objs, gas_budget),
+                Self::V1(status) => status.check_gas_balance(gas_objs, gas_budget),
             }
         }
     }
